+++ src/components/templates/ProductTemplate/ProductTemplateNavigation/ProductTemplateNavigation.tsx
@@ -1,12 +1,11 @@
 import React, { ReactNode } from 'react';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { faChevronLeft } from '@fortawesome/free-solid-svg-icons';
 
 import { Text } from '../../../..';
 import Icon from '../../../atoms/Icon/Icon';
 import Link from '../../../atoms/Link/Link';
-import { minMedia } from '../../../../helpers/responsiveness';
-import { colors } from '../../../../constants/colors';
+import { colors, grid } from '../../../../constants';
 
 interface ProductTemplateNavigationProps {
   prevStep: string | ReactNode;
@@ -19,15 +18,13 @@ const StyledLink = styled(Link)`
 `;
 
 const Wrapper = styled.div`
-  ${minMedia.desktop`
-    ${css`
-      position: absolute;
-      top: 0;
-      bottom: 0;
-      left: 0;
-      right: 0;
-    `}
-  `}
+  @media (min-width: ${grid.breakpoints.m}px) {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+  }
 `;
 
 const BackText = styled(Text).attrs({ color: colors.greyDark })`
@@ -36,7 +33,7 @@ const BackText = styled(Text).attrs({ color: colors.greyDark })`
 
 export function ProductTemplateNavigation({ prevStep }: ProductTemplateNavigationProps) {
   return (
-    <Wrapper className="my-4 l:mt-0" data-automation="ZA.ProductTemplateNavigation">
+    <Wrapper className="my-4 m:mt-0" data-automation="ZA.ProductTemplateNavigation">
       {typeof prevStep === 'string' ? (
         <StyledLink href={prevStep} aria-label="Back">
           <Icon variant={faChevronLeft} color={colors.greyDark} size="xs" className="mr-2" />