+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+### [4.26.2](https://github.com/zopaUK/react-components/compare/v4.26.1...v4.26.2) (2020-10-15)
+
+
+### Bug Fixes
+
+* **numbertext:** add titleClassName prop ([ff37e33](https://github.com/zopaUK/react-components/commit/ff37e33f7250e78e4abeeb280ad248fcf64f772c))
+
+### [4.26.1](https://github.com/zopaUK/react-components/compare/v4.26.0...v4.26.1) (2020-10-15)
+
+
+### Bug Fixes
+
+* **product-template:** adjust spacing and styles ([a7ed25d](https://github.com/zopaUK/react-components/commit/a7ed25db5adfee6601e86d0d71a24ddeaf54af9b))
+* **product-template:** fix styles ([67fc28a](https://github.com/zopaUK/react-components/commit/67fc28a2c77599af90224c478383fa8f7dbe8a18))
+* **product-template:** fix styles ([cfab275](https://github.com/zopaUK/react-components/commit/cfab2757a7e3e814ca4181bfa6ed516c37ca9e18))
+
+## [4.26.0](https://github.com/zopaUK/react-components/compare/v4.25.0...v4.26.0) (2020-10-14)
+
+
+### Features
+
+* **icon:** fix icon typings ([cfea2d5](https://github.com/zopaUK/react-components/commit/cfea2d590ab3f0b37203f36458f4ec49c790b3e2))
+
+## [4.25.0](https://github.com/zopaUK/react-components/compare/v4.24.2...v4.25.0) (2020-10-14)
+
+
+### Features
+
+* **footer:** tests update ([4009c0d](https://github.com/zopaUK/react-components/commit/4009c0d4d251eb60349aad43e9d4265a3c254228))
+* **footer:** update footer link ([5c36aa7](https://github.com/zopaUK/react-components/commit/5c36aa757657963524359ee70bbe676663f8885e))
+
+### [4.24.2](https://github.com/zopaUK/react-components/compare/v4.24.1...v4.24.2) (2020-10-14)
+
+
+### Bug Fixes
+
+* **number text:** handle negative numbers ([70b7dd0](https://github.com/zopaUK/react-components/commit/70b7dd042c5addbab24cc6fd0292c187e7d30bd1))
+
 ### [4.24.1](https://github.com/zopaUK/react-components/compare/v4.24.0...v4.24.1) (2020-10-12)